Changed LedgerConfiguration creation
