+++ src/main/kotlin/org/epirus/local/ledger/LedgerConfiguration.kt
@@ -13,54 +13,8 @@
 package org.epirus.local.ledger
 
 import org.epirus.local.cli.Account
-import org.web3j.crypto.Credentials
-import org.web3j.crypto.Keys
-import java.io.File
 
-class LedgerConfiguration(
+data class LedgerConfiguration(
     var genesis: String? = null,
-    val directory: String? = null,
     var accounts: List<Account>? = null
-) {
-    init {
-        if (genesis == null) {
-            accounts = generateAccounts()
-            genesis = createGenesis()
-        }
-    }
-
-    fun createGenesis(): String {
-        val genesis = File(directory?.removeSuffix("/") + "/genesis.json")
-        genesis.createNewFile()
-        var json = """{
-                  "comment": "Automatically generated genesis. DO NOT MODIFY!",
-                  "config": {
-                    "chainId": 999,
-                    "constantinopleFixBlock": 0,
-                    "ethash": {
-                      "fixeddifficulty": 100
-                    }
-                  },
-                  "gasLimit": "0x1000000",
-                  "difficulty": "0x10000",
-                  "alloc": {""".trimIndent()
-
-        accounts?.forEach { t -> json += "\"" + t.address + """":{"balance" : "0x56BC75E2D63100000"},""" }
-
-        json = json.removeSuffix(",") + "}}"
-        genesis.writeText(json)
-        return genesis.absolutePath
-    }
-
-    fun generateAccounts(): List<Account> {
-        val accounts = mutableListOf<Account>()
-
-        (1..10).forEach { _ ->
-            val ecKeyPair = Keys.createEcKeyPair()
-
-            val creds = Credentials.create(ecKeyPair.privateKey.toString(16))
-            accounts.add(Account(creds.address, ecKeyPair.privateKey.toString(16)))
-        }
-        return accounts
-    }
-}
+)
